ml model prediction/ml model prediction/requirements_flask.txt
flask
flask-sqlalchemy
pandas
numpy
scikit-learn
